Refactoring carrying

+++ Assets/FX/Scripts/Hexagon/BaseHexagon.cs
@@ -9,7 +9,7 @@ public class BaseHexagon : MonoBehaviour, Hexagon
     public Vector3 Position { get => FloorHexagon.Position; set => FloorHexagon.Position = value; }
     public HEX_TYPE Type { get => FloorHexagon.Type; set => FloorHexagon.Type = value; }
     public FloorHexagon FloorHexagon { get; set; }
-    public int Storage { get; set; }
+    public float Storage { get; set; }
 
     public static BaseHexagon CreateHexagon(FloorHexagon parent, BaseHexagon baseHexPrefab)
     {
