+++ Assets/FX/Scripts/Hexagon/CollectingHexagon.cs
@@ -8,8 +8,8 @@ public class CollectingHexagon : MonoBehaviour, Hexagon
     public string Id { get => FloorHexagon.Id; set => FloorHexagon.Id = value; }
     public Vector3 Position { get => FloorHexagon.Position; set => FloorHexagon.Position = value; }
     public HEX_TYPE Type { get => FloorHexagon.Type; set => FloorHexagon.Type = value; }
-    public int Quantity { get; set; }
-    public static int MaxQuantity = 1000;
+    public float Quantity { get; set; }
+    public static float MaxQuantity = 1000;
     public FloorHexagon FloorHexagon { get; set; }
     public List<Worker> Carriers { get; set; }
